Adapt to US friend's dataset.

+++ Ground-Fusion++/lio/src/tools/point_types.h
@@ -140,7 +140,7 @@ namespace ouster_ros
         float intensity;
         uint32_t t;
         uint16_t reflectivity;
-        uint8_t ring;
+        uint16_t ring;
         uint16_t ambient;
         uint32_t range;
         EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -149,16 +149,16 @@ namespace ouster_ros
 
 // clang-format off
 POINT_CLOUD_REGISTER_POINT_STRUCT(ouster_ros::Point,
-                                  (float, x, x)
-                                      (float, y, y)
-                                      (float, z, z)
-                                      (float, intensity, intensity)
-                                      // use std::uint32_t to avoid conflicting with pcl::uint32_t
-                                      (std::uint32_t, t, t)
-                                      (std::uint16_t, reflectivity, reflectivity)
-                                      (std::uint8_t, ring, ring)
-                                      (std::uint16_t, ambient, ambient)
-                                      (std::uint32_t, range, range)
+                                (float, x, x)
+                                    (float, y, y)
+                                    (float, z, z)
+                                    (float, intensity, intensity)
+                                    // use std::uint32_t to avoid conflicting with pcl::uint32_t
+                                    (std::uint32_t, t, t)
+                                    (std::uint16_t, reflectivity, reflectivity)
+                                    (std::uint16_t, ring, ring)
+                                    (std::uint16_t, ambient, ambient)
+                                    (std::uint32_t, range, range)
 )
 // clang-format on
